Robustness for GroovyCompilationUnitTests#testGetModuleNode_7(). Test ModuleNodeMapper cache after each test tear down method.Basic refactoring and clean up.
